app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
fastapi==0.112.1
fireworks-ai==0.15.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
idna==3.8
numpy==2.1.0
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pytesseract==0.3.13
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
